% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{annualAnomaly}
\alias{annualAnomaly}
\title{Calculate annual anomaly values}
\usage{
annualAnomaly(d, normalizedAnomaly = TRUE)
}
\arguments{
\item{d}{a data.frame containing year, month, and at least anomaly, and optionally normalizedAnomaly}

\item{normalizedAnomaly}{a logical value indicating whether or not to calculate the
normalized annual anomaly as well, the default is \code{TRUE}}
}
\value{
a data.frame with year, anomaly, and optionally normalizedAnomaly
}
\description{
This function calculates annual anomaly values from provided
monthly anomaly values.
}
\author{
Chantelle Layton
}
