% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{annualMeanCutSplit}
\alias{annualMeanCutSplit}
\title{Calculate annual mean}
\usage{
annualMeanCutSplit(x, y, minN = NULL)
}
\arguments{
\item{x}{the times of observations in POSIXct format.}

\item{y}{the observations}

\item{minN}{the minimum number of data points}
}
\value{
A list containing the resulting annual mean value of x, y, the number of data points,
within that year, n.
}
\description{
This function calculates the annual mean value from data that has a temporal
resolution greater than a year.
}
\author{
Chantelle Layton
}
