% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anomalyColors}
\alias{anomalyColors}
\title{Anomaly colours}
\format{
a list of \code{colors} and \code{breaks}
}
\usage{
anomalyColors
}
\description{
Color definitions for anomaly plots. Note that the breaks are one unit longer than the color
map. This is fine when supplying it to oce's colorMap function
}
\keyword{datasets}
