% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{approxPressureCtd}
\alias{approxPressureCtd}
\title{Linearly interpolate a CTD object in pressure space.}
\usage{
approxPressureCtd(x, bin, trimBin = FALSE)
}
\arguments{
\item{x}{a ctd object}

\item{bin}{a vector of numerical values of the center of the bin}

\item{trimBin}{a logical value indicating if the supplied bin values should be trimmed to the
maximum pressure, default is FALSE.}
}
\value{
a ctd object that has the same metadata and processing log as the supplied ctd object,
but with bin-averaged data.
}
\description{
Vertically average and CTD object through pressure by defined bins and tolerances.
}
\author{
Chantelle Layton
}
