% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areaClimatologies1980to2010}
\alias{areaClimatologies1980to2010}
\title{Climate area climatologies}
\format{
a list of data frames that are named by their area, for
example, 05 is an element in a list named \code{area05}, and so on for
the areas defined above. Each element in the list has a data frame
which has \code{depth}, \code{month}, \code{temperature_ltm},
\code{salinity_ltm}, and \code{sigmat_ltm}.
}
\usage{
areaClimatologies1980to2010
}
\description{
Select climatologies for specified climate areas on the scotian shelf
associated with annual reporting. Those that are analyzed yearly are
\tabular{ll}{
\strong{area number} \tab \strong{area name} \cr
05 \tab Misaine Bank \cr
12 \tab Emerald Basin \cr
24 \tab Lurcher Shoal \cr
26 \tab Georges Basin\cr
28 \tab Eastern Georges Bank \cr
66 \tab Cabot Strait
}
}
\keyword{datasets}
