% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areaDepthBins}
\alias{areaDepthBins}
\title{Area standard depth bins}
\format{
a data frame with columns \code{bin} and \code{tolerance}
}
\usage{
areaDepthBins
}
\description{
Area standard depth bins and tolerances definitions for vertically averaging
}
\keyword{datasets}
