% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oax.R
\name{as.oax}
\alias{as.oax}
\title{Format CTD for input into optimal estimation routine}
\usage{
as.oax(
  x,
  var,
  longitude0,
  latitude0,
  addDayNumber = FALSE,
  onlyBottom = FALSE,
  noise = 0.1,
  utm = TRUE,
  zone = 19
)
}
\arguments{
\item{x}{a list of CTD profiles}

\item{var}{Variable name to be extracted}

\item{longitude0, latitude0}{Reference location of the point from which distance is measured for each
station if utm is FALSE, ignored otherwise}

\item{addDayNumber}{Logical indicating whether to append day number to output}

\item{onlyBottom}{Logical indicating whether to extract only bottom values}

\item{noise}{Numeric value required for input into the optimal estimation routine, default
is set to \code{0.1} which should be used.}

\item{utm}{Logical indicating whether or not to use lonlat2utm function to calculate distances,
if FALSE, geodDist is used}

\item{zone}{zone used in calculation if utm is TRUE, ignored otherwise. Default is associated
with the summer groundfish area.}
}
\description{
This function takes a list of CTD profiles and puts them into
the format that is required for the optimal estimation routine.
}
\details{
The data format required to run the optimal estimation routine requires
that data have columns of data be x-dist, y-dist, depth, dayNumber(optional), data, noise
}
\author{
Chantelle Layton
}
