% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{binArea}
\alias{binArea}
\title{Calculate the approximate area for specified bins}
\usage{
binArea(
  x,
  longitude,
  latitude,
  gridDiffLongitude,
  gridDiffLatitude,
  useArea = TRUE,
  areaLongitude,
  areaLatitude,
  useBins = TRUE,
  bins
)
}
\arguments{
\item{x}{a vector of the data that is to be used to calculate the weighted area}

\item{longitude}{a vector of the same length as x}

\item{latitude}{a vector of the same length as x}

\item{gridDiffLongitude}{a numerical value indicating the grid spacing in longitude space}

\item{gridDiffLatitude}{a numerical value indicating the grid spacing in latitude space}

\item{useArea}{a logical indicating if x should be subset within an area}

\item{areaLongitude}{a vector of longitude values defining the area, ignored if useArea = FALSE}

\item{areaLatitude}{a vector of latitude values defining the area, ignored if useArea = FALSE}

\item{useBins}{logical indicating whether to bin x}

\item{bins}{a vector defining how to bin x, ignored if useBins = FALSE}
}
\value{
a list of bin0 and bin1 which are the ranges in which data was averaged over.
binArea, which is the calculated area within each bin range, npoints, which is the
number of points included in the calculation. And finally, avg, which is the average
value within the bin.
}
\description{
Calculate the area weighted value of a variable within a polygon
}
\author{
Chantelle Layton
}
