% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brownsBankPolygon}
\alias{brownsBankPolygon}
\title{Browns bank polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
brownsBankPolygon
}
\description{
Polygon defining the bounds to look for data along the core Atlantic Zone Monitoring Program
line, browns bank. It was constructed by finding the angle between the first point, and the sixth
station, and then the sixth station to the seventh station and the line has a leg at the end. The bounding
box was calculated to be 8 km wide.
}
\keyword{datasets}
