% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculateStratificationIndex}
\alias{calculateStratificationIndex}
\title{Calculate stratification index}
\usage{
calculateStratificationIndex(ctd, depth1, depth2, debug = TRUE)
}
\arguments{
\item{ctd}{a \code{ctd} object.}

\item{depth1}{a numeric value, such that depth1 < depth2}

\item{depth2}{a numeric value, such that depth2 > depth1}

\item{debug}{a logical value indicating if debug info should be displayed}
}
\value{
a single row data frame with relevant information for easy output use
}
\description{
This function calculates the stratification index, which is
defined as the difference in density between two defined depths divided
by the difference in the two depths. The calculation completes two
iterations. The first is between the two defined depths, and then if the
calculated value is less than 0, the depth range is slighly increased
}
\author{
Chantelle Layton and Benoit Casault
}
