% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahccd.R
\name{checkCurrentTimeForClimateSummary}
\alias{checkCurrentTimeForClimateSummary}
\title{Check current time}
\usage{
checkCurrentTimeForClimateSummary(x)
}
\arguments{
\item{x}{a POSIXlt formatted time}
}
\value{
a POSIXct formatted time
}
\description{
This function checks a provided time, and returns the most recent date
available for monthly climate data.
}
\author{
Chantelle Layton
}
