% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{cilVolume}
\alias{cilVolume}
\title{Calculate cold intermediate layer thickness}
\usage{
cilVolume(T, p, longitude, latitude, dlongitude, dlatitude, Tlim = 4)
}
\arguments{
\item{T}{a vector of temperatures}

\item{p}{a vector of pressure}

\item{longitude, latitude}{the location}

\item{dlongitude, dlatitude}{the grid spacing the longitude and latitude coordinates respectively}

\item{Tlim}{cold intermediate layer temperature limit, default of 4}
}
\value{
a list with thickness with units kilometer, volume with units kilometer^3,
and minimum temperature.
}
\description{
Calculate cold intermediate layer thickness
}
\details{
This function calculates the thickness and volume of the cold intermediate layer
at a given location. The thickness at 3 different locations along the profile is calculated
in the following manner. If the surface value is below Tlim, the thickness is calculated
as (p_i + p_(i+1)) / 2. If the bottom value, (p_i - p_(i-1))/2. If interior, (p_(i+1) - p(i-1)) / 2.
}
\author{
Chantelle Layton
}
