% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climatePolygonArea}
\alias{climatePolygonArea}
\title{Climate area polygon areas}
\format{a data frame with \code{area}, \code{polyArea} and \code{region}.}
\usage{
climatePolygonArea
}
\description{
Defined climate area polygon areas for only the Maritime region (ScotiaFundy)
with units \code{km^2}. Details on how the areas were calculated is not known
at this time.
}
\keyword{datasets}
