% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climatePolygons}
\alias{climatePolygons}
\title{Climate area polygons}
\format{
a data frame with \code{area}, \code{latitude}, \code{longitude}, and \code{region}.
}
\usage{
climatePolygons
}
\description{
Defined climate area polygons for all four Fisheries and Oceans Canada
Atlantic regions, as follows
\tabular{ll}{
\strong{abbreviation} \tab \strong{region} \cr
ScotiaFundy \tab Maritimes \cr
Gulf \tab Gulf and Quebec \cr
NfldLab \tab Newfoundland and Labrador
}
}
\keyword{datasets}
