% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyClimate.R
\name{climateSummary2ahccd}
\alias{climateSummary2ahccd}
\title{Convert monthly climate data to ahccd format}
\usage{
climateSummary2ahccd(x)
}
\arguments{
\item{x}{a data.frame of monthly climate data that was manipulated after reading it in.}
}
\value{
a list that contains the same format as \link{read.ahccd}. Note that the \code{elevation} variable
is missing from the climate summary files.
}
\description{
This function takes the monthly climate data, which is read in as a data frame,
and puts it into the decided list format of the ahccd data.
}
\author{
Chantelle Layton
}
