% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyClimate.R
\name{combineAhccdAndClimateSummary}
\alias{combineAhccdAndClimateSummary}
\title{Combine ahccd and climate summary data}
\usage{
combineAhccdAndClimateSummary(ahccd, climateSummary)
}
\arguments{
\item{ahccd}{a list that has been read in using \code{read.ahccd}}

\item{climateSummary}{a list that has been converted to the ahccd format using \code{climateSummary2ahccd}}
}
\value{
a list in the same form as \code{read.ahccd}, but all metadata has been preserved. Any metadata from
the \code{climateSummary} data has \code{climate} added in front of it's name.
}
\description{
This function combines ahccd and climate summary data together into the same list.
}
\details{
There is some additional data that is read in and retained for the climate summary
data, but here only \code{meanTemperature} is retained. The climate summary metadata has \code{climateSummary}
appended to the name in the event that a comparisson is required. Ahccd data takes precedence,
so when comparing the year and month, the ahccd data is kept over the climate summary data.
}
\author{
Chantelle Layton
}
