% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.R
\name{createIcesWgohTimeserieFile}
\alias{createIcesWgohTimeserieFile}
\title{Create ICES WGOH formatted file for submission}
\usage{
createIcesWgohTimeserieFile(
  data,
  destdir,
  fileDescriptor,
  fileType,
  index,
  stationDescription,
  regionCode,
  latitude,
  longitude,
  detailedLocation,
  measurementDepth,
  samplingFrequency,
  longTermAveragingPeriod,
  longTermMean,
  longTermStandardDeviation,
  calculationMethodology,
  dataSource,
  contactName,
  notes,
  reference,
  doi,
  website
)
}
\arguments{
\item{data}{a data.frame of the data that is to be output}

\item{destdir}{a connection object or a character string indicating the directory the file should be saved.}

\item{fileDescriptor}{A character vector of length one. Indicates a brief description of the timeseries
for the filename.}

\item{fileType}{A character vector of length one. Indicates the time series file type.
Accepted options are "Annual", "Mnthly", and "Timeseries"}

\item{index}{A numeric vector of length one. It is the index as agreed by the WGOH regional group.}

\item{stationDescription}{A character vector of length one. Short station
description/name. This needs to be same across all file resolutions relating
to the same station/region (_Annual, _Mnthly, and _Timeseries) and the different
depth levels. For example, if three different timeseries relate to the same
station but different depth levels, then this needs to be the same for all
three of these timeseries.}

\item{regionCode}{A character vector of length one. It is the region code as agreed by
WGOH regional group. It is the region the timeseries is reported in.}

\item{latitude}{A numeric vector of length one. Nominal latitude position (for plotting purposes).
Max 3 decimal places.}

\item{detailedLocation}{A Well Known Text (WKT) formatted object to describe more detailed position
information (e.g. point, linestring, polygon, multilinestring, multipoint, multipolygon).}

\item{measurementDepth}{Depth or depth range related to the time series.}

\item{samplingFrequency}{A character vector of length one. Short description of how often sampling occurs.}

\item{longTermAveragingPeriod}{A character vector of length one. The climatological reference period
in the format of \code{begin}-\code{end}, preferably 1991-2020. This is only included in fileType = "Annual".}

\item{longTermMean}{A numeric vector indicating the long term mean for the defined
\code{longTermAveragingPeriod}. for each parameter in \code{data}. Temperature should have a maximum
of 2 decimal places and salinity should have a maximum of 3 decimal places.
This is only included in fileType = "Annual".}

\item{longTermStandardDeviation}{A numeric vector indicating the long term standard deviation
for the defined \code{longTermAveragingPeriod}. for each parameter in \code{data}.
Temperature should have a maximum of 2 decimal places and salinity should have
a maximum of 3 decimal places. This is only included in fileType = "Annual".}

\item{calculationMethodology}{A character vector of length one. A short description of calculation methodology.
It should provide details of the methods for spatial averaging (in horizontal space and through depth),
temporal averaging, removal of seasonality, and calculation monthly/annual means; as appropriate.
It should also include details of how the long-term mean and standard deviation were calculated.
This is only included in fileType = "Annual".}

\item{dataSource}{A character vector of length one. The organization who provides the data in the format
\verb{acronym-full name-country}.}

\item{contactName}{A character vector of length one, with contact name and email address, in the format
\code{name (email)}. Generic inboxes are allowed for groups where an individual name is less appropriate.}

\item{notes}{An optional character vector of length one. Provides additional notes relating to the time series.}

\item{reference}{An optional character vector of length one. Journal publication reference, it should be
in quotation marks to avoid comma's in the reference causing issues in the CSV format.}

\item{doi}{An optional character vector of length one. The DOI of the dataset.}

\item{website}{An optional character vector of length one. The website for the project or dataset.}

\item{lloongitude}{A numeric vector of length one. Nominal latitude position (for plotting purposes).
Max 3 decimal places.}
}
\description{
Outputs a ICES WGOH formatted file for data submission. Based off the standard format provided on 20251109.
}
\author{
Chantelle Layton
}
