% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oax.R
\name{createOaxDeckfile}
\alias{createOaxDeckfile}
\title{Create deckfile for OAX}
\usage{
createOaxDeckfile(dependent, dataFile, gridFile, which)
}
\arguments{
\item{dependent}{a character string indicating the name of the dependent
variable}

\item{dataFile}{a character string indicating the file name associated
with the input data}

\item{gridFile}{a character string indicating the file name associated with
the OAX grid}

\item{which}{a character string indicating which OAX analysis, options include
\code{winterGroundfish}, \code{summerGroundfish}, \code{snowCrab}.}
}
\description{
This function will create character strings necessary
to save in the form of a deck file for input into OAX analysis.
}
\details{
Currently, there are three different OAX analyses that are completed,
the parameter \code{which} indicates the specific program.
}
\author{
Chantelle Layton
}
