% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{densityGradientMonthlyClimatology1981to2010}
\alias{densityGradientMonthlyClimatology1981to2010}
\title{Density gradient monthly climatology 1981 to 2010}
\format{
a data.frame that has the climatology monthly average for areas 1 to 29
of the Scotian-Fundy polygons. Variables include the value at 0m and 50m for temperature,
salinity, and sigmaTheta, and the number of profiles used in the calculation. Also, the
salinity and sigmaTheta gradient, as well as the number of profiles used for the gradient calculations.
}
\usage{
densityGradientMonthlyClimatology1981to2010
}
\description{
Relevant parameters used to calculate the density gradient across the scotian shelf
}
\keyword{datasets}
