% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{densityGradientWeightedAnomaly}
\alias{densityGradientWeightedAnomaly}
\title{Density gradient weighted anomaly}
\format{
a data.frame that has year and the weighted anomaly, weighted anomaly standard deviation, and number
of regions for variables at 0m and 50m for salinity, temperature, and sigmaTheta, along with the salinity
and sigma theta gradient
}
\usage{
densityGradientWeightedAnomaly
}
\description{
Previously calculated anomaly values for variables associated with density gradient analysis.
}
\keyword{datasets}
