% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahccd.R
\name{download.ahccd}
\alias{download.ahccd}
\title{Download temperature AHCCD data files}
\usage{
download.ahccd(destdir = ".")
}
\arguments{
\item{destdir}{Optional string indicating the directory in which to store downloaded files.
If not supplied, "." is used, i.e. the data file is stored in the present working directory.}
}
\description{
This function downloads the two files associated with the Adjusted and Homogenized
Canadian Climate Data (AHCCD) mean temperature data.
}
\details{
Download the AHCCD mean temperature data and its associated station excel file.
}
\author{
Chantelle Layton
}
