% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyClimate.R
\name{download.climateSummaries}
\alias{download.climateSummaries}
\title{Download a Cache a Met Climate Summary File}
\usage{
download.climateSummaries(
  year,
  month,
  province,
  type,
  destdir = ".",
  destfile,
  force = TRUE
)
}
\arguments{
\item{year}{A number giving the year of interest.}

\item{month}{A number giving the month of interest.}

\item{province}{A string giving the province abbreviation of interest.}

\item{type}{A string indicating which type of file to download, either
\code{"xml"} for an XML file or \code{"csv"} for a CSV file.}

\item{destdir}{Optional string indicating the directory in which to store downloaded files.
If not supplied, "." is used, i.e. the data file is stored in the present working directory.}

\item{destfile}{Optional string indicating the name of the file. If not supplied, the file
name is constructed from the other parameters of the function call, so subsequent calls with
the same parameters will yield the same result, thus providing the key to the caching scheme.}

\item{force}{Logical value indicating whether to force a download, even if the file already
exists locally.}
}
\description{
Data are downloaded from the Environment and Climate Change Canada's
climate weather website and cached locally
}
\details{
The data are downloaded from
\url{http://climate.weather.gc.ca/prods_servs/cdn_climate_summary_report_e.html}
using \code{\link[utils:download.file]{utils::download.file()}}
pointed to the Environment and Climate Change Canada website (reference 1)
using queries that had to be devised by reverse-engineering, since the agency
does not provide documentation about how to construct queries. Caution: the
query format changes from time to time, so \code{\link[=download.met.climateSummaries]{download.met.climateSummaries()}} may work one
day, and fail the next.
}
\author{
Chantelle Layton
}
