% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn.R
\name{download.ghcn}
\alias{download.ghcn}
\title{Download temperature GHCN data files}
\usage{
download.ghcn(destdir = ".")
}
\arguments{
\item{destdir}{Optional string indicating the directory in which to store downloaded files.
If not supplied, "." is used, i.e. the data file is stored in the present working directory.}
}
\description{
This function downloads the two files associated with the Global Historical
Climate Network (GHNC) mean temperature data.
}
\details{
Download the GHCN mean temperature data. As of now, it only obtains the quality
controlled adjusted data, below is a table summarizing the other types of products
\tabular{ll}{
\strong{product code} \tab \strong{product description} \cr
qcu \tab quality controlled \cr
qcf \tab quality controlled, adjusted using pairewise homogeneity algorithm \cr
qfe \tab quality controlled, adjusted using pairewise homogenetiy algorithm, 1961-2010
}
}
\author{
Chantelle Layton
}
