% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iml.R
\name{download.imlCTD.erddap}
\alias{download.imlCTD.erddap}
\title{Download IML data from CIOOS ERDDAP site}
\usage{
download.imlCTD.erddap(startYear, endYear)
}
\arguments{
\item{startYear}{A numeric vector of length one indicating the start year for the data extraction.}

\item{endYear}{A numeric vector of length one indicating the end year for the data extraction.}
}
\value{
list of oce ctd objects
}
\description{
Download IML data from CIOOS ERDDAP site and create ctd objects
for each associated \code{id}.
}
\author{
Chantelle Layton
}
