% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meds.R
\name{download.meds}
\alias{download.meds}
\title{Download MEDS data}
\usage{
download.meds(year, month, destdir = ".")
}
\arguments{
\item{year}{Number giving the year of interest}

\item{month}{Optional number indicating the month}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory. Also, the directory will be created
if not already done so.}
}
\description{
Download data hosted by Fisheries and Oceans Canada
(DFO) Marine Environmental Data Section (MEDS).
}
\details{
This function downloads data hosted by Fisheries and Oceans Canada
(DFO) Marine Environmental Data Section (MEDS) via an ftp connection.
Users have the option of supplying the year and or month. Note that the year
and month specified does not necessarily mean that the data contained in
those files was aquired during the year and month in question. The data file
will contain all the data received by MEDS during that year and month. For each
requested year and or month will download two files, a \code{.hdr} and a \code{.prf}
file. Both are required when reading the data as the \code{.hdr} file contains
some basic metadata, and \code{.prf} contains all of the data.
}
\author{
Chantelle Layton
}
