% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nao.R
\name{download.nao.hurrell}
\alias{download.nao.hurrell}
\title{Download and Cache the Hurrell Principal Component North Atlantic Oscillation data}
\usage{
download.nao.hurrell(destdir = ".", destfile = NULL, which = "DJFM")
}
\arguments{
\item{destdir}{Optional string indicating the directory in which to store downloaded files.
If not supplied, "." is used, i.e. the data file is stored in the present working directory.}

\item{destfile}{Optional string indicating the name of the file. If not supplied, the file
name as provided on the website.}

\item{which}{Required string indicating which flavour of the NAO should be downloaded, current options
are 'annual', 'monthly', 'DJF', 'DJFM', 'MAM', 'JJA', 'SON', see the website for details. Default
is 'DJFM'.}
}
\description{
Data are downloaded from the National Science Foundation (NSF) National Center for Atmospheric
Research (NCAR) Climate Data Guide website.
}
\details{
The data are downloaded from
\url{'https://climatedataguide.ucar.edu/climate-data/hurrell-north-atlantic-oscillation-nao-index-pc-based'}
using \code{\link[utils:download.file]{utils::download.file()}}
pointed to the National Science Foundation (NSF) National Center for Atmospheric
Research (NCAR) Climate Data Guide website. Caution: the
location that these data are held could potentially vary, so it could work one
day, and fail the next.
}
