% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nao.R
\name{download.nao}
\alias{download.nao}
\title{Download and Cache the North Atlantic Oscillation data}
\usage{
download.nao(destdir = ".", destfile = NULL)
}
\arguments{
\item{destdir}{Optional string indicating the directory in which to store downloaded files.
If not supplied, "." is used, i.e. the data file is stored in the present working directory.}

\item{destfile}{Optional string indicating the name of the file. If not supplied, the file
name as provided on the website.}
}
\description{
Data are downloaded from the National Centers for Environmental Information
National Oceanic and Atmospheric Administration website.
}
\details{
The data are downloaded from
\url{'https://www.ncdc.noaa.gov/teleconnections/nao/data.csv'}
using \code{\link[utils:download.file]{utils::download.file()}}
pointed to the National Centers for Environmental Information National
Oceanic and Atmospheric Administration website. Caution: the
location that these data are held could potentially vary, so it could work one
day, and fail the next.
}
