% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{download.nefsc.erddap}
\alias{download.nefsc.erddap}
\title{Download NODC NEFSC data from erddap server}
\usage{
download.nefsc.erddap(startDate, endDate, destdir = ".")
}
\arguments{
\item{startDate}{A character string in the format \code{yyyy-mm-dd} identifying the start of the desired
timeseries beginning at 00:00UTC}

\item{endDate}{A character string in the format \code{yyyy-mm-dd} identifying the end of the desired
timeseries ending at 23:59UTC}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory. Also, the directory will be created
if not already done so.}
}
\value{
a character string of the filename.
}
\description{
Download data hosted by the National
Oceanic and Atmospheric Administration (NOAA) Northeast Fisheries
Science Center (NEFSC).
}
\details{
This function downloads data hosted by the National
Oceanic and Atmospheric Administration (NOAA) Northeast Fisheries
Science Center (NEFSC) from the ERDDAP server. Saves a \code{.nc} file in
supplied \code{destdir}.
}
\author{
Chantelle Layton
}
