% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{download.nefsc}
\alias{download.nefsc}
\title{Download NODC NEFSC data}
\usage{
download.nefsc(year, ship, destdir = ".")
}
\arguments{
\item{year}{Optional number giving the year of interest}

\item{ship}{Optional string indicating the ship of interest}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory. Also, if the directory will be created
if not already done so.}
}
\description{
Download data hosted by the National
Oceanic and Atmospheric Administration (NOAA) Northeast Fisheries
Science Center (NEFSC).
}
\details{
This function downloads data hosted by the National
Oceanic and Atmospheric Administration (NOAA) Northeast Fisheries
Science Center (NEFSC) via an ftp connection. Users have the option of supplying
either the year and or ship, but one must be supplied.
}
\section{Valid ship abbreviations with full name}{

\tabular{ll}{
\strong{abbreviation} \tab \strong{full name} \cr
aj   \tab \cr
alb  \tab \cr
am   \tab \cr
arm  \tab RV Neil Armstrong \cr
c1   \tab \cr
ch   \tab \cr
del  \tab FSV Delaware II \cr
edl  \tab \cr
eg   \tab \cr
egg  \tab \cr
en   \tab E/V Endeavor \cr
end  \tab \cr
ex   \tab Okeanos Explorer \cr
ey   \tab Eagle Eye II \cr
gu   \tab Gordon Gunter \cr
hb   \tab FSV Henry B. Bigelow \cr
is   \tab \cr
kat  \tab \cr
nob  \tab \cr
np   \tab \cr
oce  \tab \cr
ore  \tab \cr
pc   \tab FSV Pisces \cr
pe   \tab \cr
rel  \tab \cr
s1   \tab R/V Hugh R. Sharp \cr
sj   \tab \cr
unk  \tab
}
}

\author{
Chantelle Layton
}
