% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neracoos.R
\name{download.neracoos.buoy}
\alias{download.neracoos.buoy}
\title{Download NERACOOS buoy CTD data}
\usage{
download.neracoos.buoy(buoyName, startDate, endDate, destdir = ".")
}
\arguments{
\item{buoyName}{A character string identifying the name of the buoy. As of now 20210104 a good
definition of all the buoy data currently available has not been identified. Although, the old
method of downloading the data has a list of buoy names with a more descriptive name,
http://www.neracoos.org/datatools/historical/graphing_download.}

\item{startDate}{A character string in the format \code{yyyy-mm-dd} identifying the start of the desired
timeseries beginning at 00:00UTC}

\item{endDate}{A character string in the format \code{yyyy-mm-dd} indentifying the end of the desired
timeseries ending at 23:59UTC}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory. Also, the directory will be created
if not already done so.}
}
\description{
Download hourly buoy CTD data hosted by Northeastern Regional Association
of Coastal Ocean Observing Systems (NERACOOS) from the ERDDAP server.
}
\details{
This function downloads data from the NERACOOS ERDDAP server via a ftp connection.
It has been specifically written to download CTD data from their buoy sites. Data chosen to download
includes all variables, including QC flags. The file is downloaded as a netcdf file and is output with the following
naming convention, \code{buoyName_sbe37_all_startDate_endDate.nc} , which uses the supplied parameters.
}
\author{
Chantelle Layton
}
