% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabass.R
\name{download.seabass.glider}
\alias{download.seabass.glider}
\title{Download SeaBASS Glider data}
\usage{
download.seabass.glider(destdir = ".")
}
\arguments{
\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory. Also, if the directory will be created
if not already done so.}
}
\description{
Download glider data that is hosted on the SeaWiFS Bio-optical
Archive and Storage System (SeaBASS) that is maintained by the
NASA Ocean Biology Processing Group (OBPG).
}
\details{
This function downloads glider data that is hosted on the SeaWiFS Bio-optical
Archive and Storage System (SeaBASS) site that is maintained by the
NASA Ocean Biology Processing Group (OBPG). Since there is no ftp site
to download data from, the website is interrograted to obtain a list
of glider files, and the files are simply downloaded from the website.
Since there is no descriptive information on the time range that each
file covers, all files are downloaded and it is up to the user to identify
which data is of use.
}
