% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{findBottomDepthIndex}
\alias{findBottomDepthIndex}
\title{Find bottom depth index for given list of longitude, latitude, and depth}
\usage{
findBottomDepthIndex(longitude, latitude, depth)
}
\arguments{
\item{longitude}{a numeric vector of longitudes}

\item{latitude}{a numeric vector of latitudes}

\item{depth}{a numeric vector of depths}
}
\value{
a logical vector of TRUE/FALSE, TRUE indicating the bottom depth value for the unique
longitude,latitude pair
}
\description{
This function takes in unlisted longitude, latitude, and depth values and finds
the index at which the maximum depth for each unique longitude,latitude pair occurs
}
\author{
Chantelle Layton
}
