% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{findValidSnowCrabGridPoints}
\alias{findValidSnowCrabGridPoints}
\title{Find valid snow crab oax grid points}
\usage{
findValidSnowCrabGridPoints(
  longitude,
  latitude,
  pressure = NULL,
  checkNthl = TRUE,
  checkIgnore = TRUE,
  checkPressure = FALSE,
  pressureMinimum = 20,
  pressureMaximum = 200
)
}
\arguments{
\item{longitude, latitude}{an optional numeric vector of longitudes and latitudes,
note that both must be supplied if \code{checkNthl} or \code{checkIgnore} are \code{TRUE}.}

\item{pressure}{an optional numeric vector of pressure, ignored if
\code{checkPressure} is \code{FALSE}.}

\item{checkNthl}{a logical value indicating whether or not to check points in
western Northumberland, default is \code{TRUE}.}

\item{checkIgnore}{a logical value indicating whether or not to check for points
that have been identified to be ignored.}

\item{checkPressure}{a logical value indicating whether or not to check points
based on the pressure, default is \code{FALSE}.}

\item{pressureMinimum}{a numeric value indicating the minimum pressure value
for valid points, default is 20. If \code{checkPressure} is \code{FALSE}, then ignored.}

\item{pressureMaximum}{a numeric value indicating the maximum pressure value
for valid points, default is 200. If \code{checkPressure} is \code{FALSE}, then ignored.}
}
\value{
a vector of logical values, TRUE indicating a point that should be included
in analysis.
}
\description{
This function identifies points that should be included
when calculating various metrics using the oax results for snow crab analysis.
}
\details{
The oax snow crab analysis grids have varied over the years. When
a new one was developed, for continuity to past results, various points
are ommited from the analysis. See the source code for more details,
but in general, there are a handful of points that are omitted from the
western Northumberland strait, various ones to be ignored along the coast,
and points where the bottom depth of a grid point that is less than 20m and
greater than 200 m.
}
\author{
Chantelle Layton
}
