% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getData}
\alias{getData}
\title{Get data translation}
\usage{
getData(item)
}
\arguments{
\item{item}{A character string the desired label.}
}
\description{
This function serves as a interface to obtain translations for a number
of potential data fields that will be needed when saving data products.
}
\author{
Chantelle Layton
}
