% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getLabel}
\alias{getLabel}
\title{Provide names for various labels}
\usage{
getLabel(item, sep = "")
}
\arguments{
\item{item}{A character string the desired label.}

\item{sep}{An optional character string inserted between the unit and the unit bracket that encloses it.
If not provided, the default \code{oceUnitSep} is used.}
}
\description{
This function provides various labels.
}
\author{
Chantelle Layton
}
