% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getLocationName}
\alias{getLocationName}
\title{Provide labels for locations.}
\usage{
getLocationName(item)
}
\arguments{
\item{item}{A character string the desired location.}
}
\description{
This function provides labels for locations in the
Atlantic region.
}
\author{
Chantelle Layton
}
