% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata translation}
\usage{
getMetadata(item)
}
\arguments{
\item{item}{A character string the desired label.}
}
\description{
This function serves as a interface to obtain translations for a number
of potential meta data fields that will be needed when saving data products.
}
\author{
Chantelle Layton
}
