% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getNumber}
\alias{getNumber}
\title{Provide written value for numbers up to 10.}
\usage{
getNumber(x)
}
\arguments{
\item{x}{A single numerical value for the desired number.}
}
\description{
This function provides the written number, for numbers up to 10.
}
\author{
Chantelle Layton
}
