% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getOrdinal}
\alias{getOrdinal}
\title{Provide ordinal value for numbers up to 10.}
\usage{
getOrdinal(x)
}
\arguments{
\item{x}{A single numerical value for the desired number.}
}
\description{
This function provides the ordinal word, with the article ('the'), for numbers up to 10.
}
\author{
Chantelle Layton
}
