% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getShipnameLong}
\alias{getShipnameLong}
\title{Get the long name of a ship abbreviation}
\usage{
getShipnameLong(x)
}
\arguments{
\item{x}{a character string indicating the abbreviation of a ship}
}
\description{
This function takes an abbreviation of a ship and returns the long name.
}
\author{
Chantelle Layton
}
