% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getTransectAngle}
\alias{getTransectAngle}
\title{Get transect angle}
\usage{
getTransectAngle(longitude, latitude)
}
\arguments{
\item{longitude}{vector of longitudes}

\item{latitude}{vector of latitudes}
}
\value{
The angle from 0 degrees pointing east, as well as the neutral regression
output.
}
\description{
Get transect angle
}
\details{
This function calculates a neutral regression on
geographical coordinates and outputs the resulting angle from
the regression.
}
\author{
Chantelle Layton
}
