% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getUnit}
\alias{getUnit}
\title{Get unit translation}
\usage{
getUnit(item)
}
\arguments{
\item{item}{A character string the desired label.}
}
\description{
This function serves as a interface to obtain translations for a number
of potential units needed when plotting.
}
\author{
Chantelle Layton
}
