% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halifaxExtendedPolygon}
\alias{halifaxExtendedPolygon}
\title{Halifax extended polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
halifaxExtendedPolygon
}
\description{
Polygon defining the bounds to look for data along the core Atlantic Zone Offshore Monitoring Program
line, Halifax extended. It was constructed by finding the angle between the first point and the last point
of the line, and then calculating the bounding box which is 8 km wide. As of now, 20211003, it only includes
HL_08 to HL_12.
}
\keyword{datasets}
