% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halifaxExtendedStationPolygons}
\alias{halifaxExtendedStationPolygons}
\title{Halifax station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
halifaxExtendedStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the core Atlantic Zone Offshore Monitoring Program
line, Halifax extended. Defining bounds are determined by station spacing. If the spacing is less than 8km,
then it is modified to half of the distance between the adjacent spacing. Note that this mean that the spacing
on either side can be uneven. As of now, 20211003, it only includes HL_08 to HL_12.
}
\keyword{datasets}
