% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahccd.R
\name{identifyMissingAhccd}
\alias{identifyMissingAhccd}
\title{Find missing air temperature data}
\usage{
identifyMissingAhccd(ahccd)
}
\arguments{
\item{ahccd}{a list containing data that is read in using \code{read.ahcdd}}
}
\value{
a data.frame of year, month, province, stationId, stationName
}
\description{
This function identifies missing monthly data from the last date to
the present date
}
