% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn.R
\name{join.ghcn}
\alias{join.ghcn}
\title{Join GHCN station data}
\usage{
join.ghcn(ghcn1, ghcn2, minYear = NULL)
}
\arguments{
\item{ghcn1}{a list that was output in using \code{read.ghcn}, this is the primary station}

\item{ghcn2}{a list that was output using \code{read.ghcn}, this is the secondary station}

\item{minYear}{a numeric value indicating the minimum year to subset the data, default is \code{NULL}
to retain the entire time series.}
}
\description{
This function will join the data of two sites that might need to be combined
due to temporal coverage.
}
\author{
Chantelle Layton
}
