% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laurentianChannelMouthStationPolygons}
\alias{laurentianChannelMouthStationPolygons}
\title{Laurentian Channel Mouth station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
laurentianChannelMouthStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the core Atlantic Zone Monitoring Program
line, Laurentian Channel Mouth. Note that the bounding box around each station is 4km due to station spacing.
}
\keyword{datasets}
