% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{makeCruiseMapTitle}
\alias{makeCruiseMapTitle}
\title{Make cruise map title}
\usage{
makeCruiseMapTitle(ctd)
}
\arguments{
\item{ctd}{a list of ctd stations}
}
\description{
This function constructs a nice title that includes the
mission number, with the long ship name, the date ranges, and
the number of stations.
}
\author{
Chantelle Layton
}
