% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{makeSectionPlotTitle}
\alias{makeSectionPlotTitle}
\title{Make section plot title}
\usage{
makeSectionPlotTitle(ctd)
}
\arguments{
\item{ctd}{a list of CTD objects}
}
\description{
This function uses CTD for a given transect to make a nice title that
includes the transect name and the date range in which measurements were obtained.
}
\author{
Chantelle Layton
}
