% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahccd.R
\name{mapAhccdStationHeader}
\alias{mapAhccdStationHeader}
\title{Map AHCCD station column names}
\usage{
mapAhccdStationHeader(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
In efforts to attempt to infer column names and put them in a more useable form
for useage.
}
\author{
Chantelle Layton
}
