% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{monthlyMeanCutSplit}
\alias{monthlyMeanCutSplit}
\title{Calculate monthly mean}
\usage{
monthlyMeanCutSplit(x, y, minN = NULL)
}
\arguments{
\item{x}{the times of observations in POSIXct format.}

\item{y}{the observations}

\item{minN}{the minimum number of data points}
}
\value{
A list containing the resulting monthly mean values of x, y, the number of data points,
within that month, n.
}
\description{
This function calculates the monthly mean values from data that has a temporal
resolution greater than a month.
}
\author{
Chantelle Layton
}
