% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{monthlyStandardDeviation}
\alias{monthlyStandardDeviation}
\title{Calculate monthly climatology standard deviation}
\usage{
monthlyStandardDeviation(d, climatologyYears)
}
\arguments{
\item{d}{a data.frame containing at least a column named month and one other variable}

\item{climatologyYears}{a vector of length two indicating the range of years}
}
\value{
the results of the aggregate function, a data.frame with columns \code{month} and \code{temperature}
}
\description{
This function calculates the monthly climatology standard deviation.
}
\details{
Monthly climatology standard deviation is calculated by taking the values within the
defined climatology range and finding the standard deviation for each month.
}
\author{
Chantelle Layton
}
