% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nafoZones}
\alias{nafoZones}
\title{NAFO zones for Maritimes regions}
\format{a list for each NAFO zone, where each zone has \code{zoneName}, \code{longitude}, and
\code{latitude}}
\usage{
nafoZones
}
\description{
NAFO zone definitions that have been subset to those only concerned with in the Maritimes
region. The shapefile was downloaded from the NAFO website.
}
\keyword{datasets}
