% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{northeastChannelStationPolygons}
\alias{northeastChannelStationPolygons}
\title{North east channel station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
northeastChannelStationPolygons
}
\description{
Polygons, a 4km box, defining bounds to look for data at each station along the core Atlantic Zone Monitoring Program
line, north east channel.
}
\keyword{datasets}
