% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{paletteCalculations}
\alias{paletteCalculations}
\title{palette calculations}
\usage{
paletteCalculations(
  separation = par("cin")[2]/2 + 0.1 - 0.05 - 0.0125,
  width = par("cin")[2] + 0.2 - 0.05 - 0.0125,
  pos = 4,
  zlab,
  maidiff = c(0, 0, 0, 0),
  debug = getOption("oceDebug")
)
}
\arguments{
\item{separation}{a numeric value}

\item{width}{a numeric value}

\item{pos}{a numeric value indicating the side}

\item{zlab}{a character}

\item{maidiff}{a numeric value}

\item{debug}{a numeric value}
}
\description{
palette calculations
}
