% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAnnualAnomaly}
\alias{plotAnnualAnomaly}
\title{Plot annual time series}
\usage{
plotAnnualAnomaly(
  x,
  y,
  ysc,
  xlim,
  ylim,
  xlab = TRUE,
  climatologyYears,
  ylabel = TRUE,
  plotSd = TRUE,
  yaxs = TRUE,
  plotPoints = TRUE,
  plotRunningAvg = TRUE,
  plotLmTrend = FALSE,
  lmResults,
  plotClimatologyMean = FALSE,
  climatologyMean,
  plotScorecard = FALSE,
  scorecardLines = TRUE,
  scorecardLabels = TRUE
)
}
\arguments{
\item{x}{a vector indicating the times of observations}

\item{y}{a vector indicating the observations}

\item{ysc}{a vector indicating observations used for scorecard, ignored if \code{plotScorecard = FALSE}}

\item{xlim}{Optional limit for x axis}

\item{ylim}{Optional limit for y axis}

\item{xlab}{Logical indicating whether to label the x axis}

\item{climatologyYears}{Vector of length two indicating the climatology years, ignored if
\code{plotSd = FALSE}.}

\item{ylabel}{Logical indicating whether to label the y axis}

\item{plotSd}{Logical indicating whether to add +/- 0.5 standard deviation lines on plot of
the annual anomaly over the defined climatology period.}

\item{yaxs}{Logical indicating whether to have a labelled y-axis}

\item{plotPoints}{Logical indicating whether or not to plot points on top of line}

\item{plotRunningAvg}{Logical indicating whether or not to plot a 5 year running average}

\item{plotLmTrend}{Logical indicating whether or not to print out results of a linear regression
on the plot}

\item{lmResults}{A \code{lm} format object, if \code{plotLmTrend = FALSE}, ignored.}

\item{plotClimatologyMean}{Logical indicating whether or not to print out the climatology mean.}

\item{climatologyMean}{A numeric value, if \code{plotClimatologyMean = FALSE}, ignored.}

\item{plotScorecard}{Logical indicating whether or not to plot scorecard on side = 1.}

\item{scorecardLines}{Logical indicating whether or not to put vertical lines on scorecard.}

\item{scorecardLabels}{Logical indicating whether or not to print \code{ysc} values in scorecard.}
}
\description{
Creates a time-series of annual time series in the format that is used
for
}
\details{
The current format of the figure is for the 2019 research document. Any changes in the future
will be reflected in the code with comments, and here.

\itemize{
\item{2019 - Annual anomalies are plotted with a grey line and a 5-year running mean plotted over top in black.
Thick dashed horizonal lines of the standard deviation of the annual climatology are added. Code has
been written to ensure that at least one century is labelled.}
}
}
\author{
Chantelle Layton
}
