% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMonthlyBar}
\alias{plotMonthlyBar}
\title{Plot monthly bar plots}
\usage{
plotMonthlyBar(
  df,
  ylim,
  years,
  plotYear,
  yearLabel = FALSE,
  yearLabelSide,
  drawPalette = TRUE,
  xaxt = TRUE,
  ylabel = TRUE,
  mar
)
}
\arguments{
\item{df}{a data.frame containing at least year, month, anomaly, and normalizedAnomaly.}

\item{ylim}{a vector of length 2 indicating the range for the y-axis}

\item{years}{vector of length two indicating the range of years for climatology}

\item{plotYear}{a vector indicating which years to plot. Can be a vector of length one, indicating
to plot only one year, or of length two, indicating a range of years to plot}

\item{yearLabel}{logical expression indicating whether or not to label the plotYears}

\item{yearLabelSide}{which side the yearLabel should be placed, (1 = bottom, 3 = top)}

\item{drawPalette}{logical expression indicating whether or not to draw a palette}

\item{xaxt}{logical expression indicating whether or not to label x-axis}

\item{ylabel}{logical exptression indicating whether or not to label y-axis labels}

\item{mar}{vector of length 4 indicating the margins of the plot}
}
\description{
This plotting method for air temperature data plots bar plots
for monthly anomalies and is colour coded based on normalized monthly
anomalies
}
\author{
Chantelle Layton
}
