% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMonthlyStackedBarplot}
\alias{plotMonthlyStackedBarplot}
\title{Plot annual stacked bar plot}
\usage{
plotMonthlyStackedBarplot(d, plotAverage = TRUE, stationNames)
}
\arguments{
\item{d}{a matrix with the columns being the x-axis and the rows being items
that wish to be stacked.}

\item{plotAverage}{a logical value indicating whether or not to add the average value
from matrix d.}

\item{stationNames}{a character vector of equal length to the number of rows in d.}
}
\description{
This function will plot a stacked bar plot that has negative and
positive values. It also has the option to plot the average of values provided.
}
\author{
Chantelle Layton
}
