% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMonthlyTimeseriesWAnomalyBar}
\alias{plotMonthlyTimeseriesWAnomalyBar}
\title{Plot monthly timeseries with anomaly bar}
\usage{
plotMonthlyTimeseriesWAnomalyBar(
  xYear,
  xMonth,
  y,
  yAnomaly,
  xAnnualAnomaly,
  yAnnualAnomaly,
  xClimatology,
  yClimatology,
  sdClimatology,
  xlim,
  ylim,
  anomalyColors
)
}
\arguments{
\item{xYear}{a numeric vector of the associated year}

\item{xMonth}{a numeric vector of the associated month}

\item{y}{a numeric vector of the associated year month value}

\item{yAnomaly}{a numeric vector of the associated anomaly year month value}

\item{xAnnualAnomaly}{a numeric vector of the year for the annual anomaly}

\item{yAnnualAnomaly}{a numeric vector of the annual anomaly}

\item{xClimatology}{a numeric vector of the months for the monthly climatology}

\item{yClimatology}{a numeric vector of the climatology values for associated month}

\item{sdClimatology}{a numeric vector of the standard deviation for the associated month}

\item{xlim}{the x limits (x1, x2) of the plot}

\item{ylim}{the y limits (y1, y2) of the plot}

\item{anomalyColors}{a list that contains the colors and breaks of the desired color bar for the anomaly scorecard}
}
\description{
Creates a time series plot with the monthly anomaly and annual
anomaly scorecard below.
}
\details{
This is the plot style that is used in many of Peter Galbraith's
research documents. Was translated to R in efforts to better represent
some data.
}
\author{
Chantelle Layton
}
