% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{portsmouthStationPolygons}
\alias{portsmouthStationPolygons}
\title{Portsmouth station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
portsmouthStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the core Atlantic Zone Monitoring Program
line, Portsmouth.
}
\keyword{datasets}
