% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prince5Climatology1980to2010}
\alias{prince5Climatology1980to2010}
\title{Prince 5 climatology}
\format{
a data frame with columns \code{Month}, \code{Depth}, \code{Avg_Temperature},
\code{Std_Temperature}, \code{Years_Temperature}, \code{Avg_Salinity}, \code{Std_Salinity},
\code{Years_Salinity}, \code{Avg_Sigmat}, \code{Std_Sigmat}, \code{Years_Sigmat}.
}
\usage{
prince5Climatology1980to2010
}
\description{
Prince 5 1980 to 2010 climatology
}
\keyword{datasets}
