% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prince5depthBins}
\alias{prince5depthBins}
\title{Prince 5 standard depth bins}
\format{
a data frame with columns \code{bin} and \code{tolerance}
}
\usage{
prince5depthBins
}
\description{
Prince 5 standard depth bins and tolerance definitions for vertical averaging
}
\keyword{datasets}
