% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prince5monthlyData}
\alias{prince5monthlyData}
\title{Prince 5 monthly data}
\format{a data frame with columns \code{year}, \code{month}, \code{depth},
\code{temperature}, \code{salinity}, \code{sigmaTheta}}
\usage{
prince5monthlyData
}
\description{
Prince 5 monthly and vertically averaged data for the entire timeseries.
}
\keyword{datasets}
