% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahccd.R
\name{read.ahccd.2026}
\alias{read.ahccd.2026}
\title{Read AHCCD data, 2026 version}
\usage{
read.ahccd.2026(
  file,
  longitude = NULL,
  latitude = NULL,
  elevation = NULL,
  temperatureVar = "Adjusted"
)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load.}

\item{longitude}{optional numerical value containing longitude in decimal degrees,
positive in the eastern hemisphere.}

\item{latitude}{optional numerical value containing the latitude in decimal degrees,
positive in the northern hemisphere.}

\item{elevation}{optional numerical value containing the elevation of the station in meters.}

\item{temperatureVar}{a character vector indicating which temperature variable to output. Options include
HomT, Adjusted, Original. Default is Adjusted. See details for a description of each.}
}
\value{
named list of data for each station provided in stns
}
\description{
Read Adjusted and Homogenized Canadian Climate \code{CanHomT_mlyV3.1} Data files. Data can be downloaded from
the web site https://data-donnees.az.ec.gc.ca/data/climate/scientificknowledge/canadian-homogenized-surface-air-temperatures?lang=en . It is required that the data file be downloaded.
Unlike previous available versions of this data, an associated station file is not provided.

Flag values are not retained at the moment
}
\details{
HomT is Homogenized temperature data. Adjusted is data that has been adjusted to remove the effects of the observing
time change in 1961 or changed because of quality control. Original is origical archived data.
}
