% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahccd.R
\name{read.ahccd.stations}
\alias{read.ahccd.stations}
\title{Read AHCCD station data file}
\usage{
read.ahccd.stations(file)
}
\arguments{
\item{file}{xls file containing the station data}
}
\description{
Read in the file which defines all stations included in the Adjusted and Homogenized
Canadian Climate Data (AHCCD).
}
\details{
Read Adjusted and Homogenized Canadian Climate Data station files. Excel station can be downloaded from
the ftp site at \url{ftp://ccrp.tor.ec.gc.ca/pub/AHCCD/} . It is normally downloaded
along with the associated data file. For example, if the 'Precipitation_Stations.xls' file
is downloaded, then one of the associated precipitation files should also be downloaded, in this example case
'Adj_monthly_rain.zip' could be downloaded.
}
\author{
Chantelle Layton
}
