% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyClimate.R
\name{read.climateSummaries.xml}
\alias{read.climateSummaries.xml}
\title{Read met xml climate summary file}
\usage{
read.climateSummaries.xml(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load.}
}
\value{
a data.frame containing all of the data, except for snow and sunshine.
}
\description{
Read xml formatted climate summary files from Environment and Climate Change Canada.
Since no file format is not formally published, the function has been based off of some sample
files that were downloaded on 20200529. File formats are subject to change, therefore, this
function has the potential to break.
}
\author{
Chantelle Layton
}
