% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn.R
\name{read.ghcn}
\alias{read.ghcn}
\title{Read in ghcn data}
\usage{
read.ghcn(dataFile, metadataFile, stationId = NULL)
}
\arguments{
\item{dataFile}{a character string giving the name of the file to read, this will be the \code{.dat} file}

\item{metadataFile}{a character string giving the name of the file to read, this will be the \code{.inv} file}

\item{stationId}{a character string of the station ID to extract}
}
\value{
a list that is in similar nature to the ahccd format, it contains some metadata,
\code{stationId}, \code{stationName}, \code{longitude}, \code{latitude}, and \code{elevation}, as well as a
data that contains \code{year}, \code{month}, \code{temperature}, and \code{flag}.
}
\description{
This function reads in Global Historical Climatology Network monthly (GHCNm)
data that is downloaded using \code{download.ghcn}.
}
\author{
Chantelle Layton
}
