% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glider.R
\name{read.gliderDat}
\alias{read.gliderDat}
\title{Read .dat glider data}
\usage{
read.gliderDat(file)
}
\arguments{
\item{file}{a .dat file}
}
\value{
a list of ctd objects
}
\description{
This function is very specialized for a non-standard data format. It reads in .dat files that
are being produced, as of 20191002, for DFO gliders, as processed by Dave Hebert.
}
\details{
This function reads in the data and then puts the data into \code{oce} ctd objects for easy
use. Apart from the standard physical parameters, temperature and salinity, oxygen percent saturation
and fluoresence is also added as oxygenPercentSaturation and fluorometer respectively.
}
\author{
Chantelle Layton
}
