% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meds.R
\name{read.meds}
\alias{read.meds}
\title{Read MEDS data}
\usage{
read.meds(hdrfile, prffile, exclusionLon = NULL, exclusionLat = NULL)
}
\arguments{
\item{hdrfile}{a connection or a character string indicating the header , `.hdr``, file to read}

\item{prffile}{a connection or a character string indicating the data, `.prf``, file to read}

\item{exclusionLon}{a supplied vector of longitude coordinates to ignore, default is set to NULL,
meaning keep all data.}

\item{exclusionLat}{a supplied vector of latitude coordinates to ignore, default is set to NULL,
meaning keep all data.}
}
\value{
a list of ctd objects
}
\description{
Read in data that was downloaded, which can be done by \code{\link[=download.meds]{download.meds()}},
from Fisheries and Oceans Canada (DFO) Marine Environmental Data Section (MEDS). The function
requires that the \code{.hdr} and \code{.prf} files, which are associated with the same year and month
are supplied.
}
