% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscInstruments.R
\name{read.minilog}
\alias{read.minilog}
\title{Read Vemco minilog file}
\usage{
read.minilog(file, skipStart = NA, skipEnd = NA, skip = 8, pressure)
}
\arguments{
\item{file}{a connection or character string  of the file to read.}

\item{skipStart}{an optional numeric value indicating number of hours to skip at the beginning of the file,
default is NA, meaning keep all the data points.}

\item{skipEnd}{an optional numeric value indicating number of hours to skip at the beginning of the file,
default is NA, meaning keep all the data points.}

\item{skip}{an optional integer value indicating the number of lines of the data file to skip before
beginning to read the data, default is 8.}

\item{pressure}{a numerical value of the pressure that the instrument was moored}
}
\value{
a ctd object with \code{deploymentType = "moored"}.
}
\description{
This function reads in a Vemco minilog file.
}
\details{
Reads a Vemco minilog file assuming just date, time and temperature fields. At this time
it is not clear is this a standard formatted file. Saves no information
from the header, not the even column names. It assumes that only temperature is present. If for some reason
a conductivity or pressure sensor is added, the code for this function will need to be updated accordingly.
}
\author{
Chantelle Layton
}
