% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nao.R
\name{read.nao.hurrell}
\alias{read.nao.hurrell}
\title{Read Hurrell North Atlantic Oscillation Index file}
\usage{
read.nao.hurrell(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load.}
}
\value{
a data frame with columns named year, and value
}
\description{
This function reads in the Hurrell North Atlantic Oscillation index.
}
\author{
Chantelle Layton
}
