% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nao.R
\name{read.nao}
\alias{read.nao}
\title{Read North Atlantic Oscillation Index file}
\usage{
read.nao(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load.}
}
\value{
a data frame with columns named year, month, and value
}
\description{
This function reads in the north atlantic oscillation index
comma separated file.
}
\author{
Chantelle Layton
}
