% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{read.nefsc.erddap.nc}
\alias{read.nefsc.erddap.nc}
\title{Read NODC NEFSC Profile data}
\usage{
read.nefsc.erddap.nc(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to read.}
}
\value{
A list of \code{ctd} objects.
}
\description{
This function reads in NODC NEFSC Profile data that was downloaded using \link{download.nefsc.erddap}.
}
\author{
Chantelle Layton
}
