% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{read.nefsc}
\alias{read.nefsc}
\title{Read NEFSC hydrographic data}
\usage{
read.nefsc(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to read.}
}
\value{
a list of ctd objects
}
\description{
Read in Northeast Fisheries
Science Center (NEFSC) hydrographic
data that can be downloaded using \code{\link[=download.nefsc]{download.nefsc()}}.
}
