% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neracoos.R
\name{read.neracoos.buoy}
\alias{read.neracoos.buoy}
\title{Read NERACOOS buoy data}
\usage{
read.neracoos.buoy(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to read.}
}
\value{
A list of \code{ctd} objects, one for each depth.
}
\description{
This function reads in NERACOOS buoy data that was downloaded using \link{read.neracoos.buoy}.
}
\author{
Chantelle Layton
}
