% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabass.R
\name{read.seabass.glider}
\alias{read.seabass.glider}
\title{Read SeaBASS glider data}
\usage{
read.seabass.glider(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to read.}
}
\value{
a ctd object
}
\description{
Read in SeaBASS glider data files that were downloaded
using \code{\link[=download.seabass.glider]{download.seabass.glider()}}.
}
\details{
This function reads in glider data files that were downloaded
using \code{\link[=download.seabass.glider]{download.seabass.glider()}}. As of now, since there is no formal
way of handling glider data, the data is returned as an ctd object which
can be further used to obtain individual profiles.
}
