% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel2.R
\name{read.sealevel2}
\alias{read.sealevel2}
\title{Read sea level data}
\usage{
read.sealevel2(file, combineData = TRUE)
}
\arguments{
\item{file}{a connection or a character string or vector giving the name(s) of the
file(s) to load.}

\item{combineData}{a logical value indicating whether or not to combine data
into a single time-series if the length of files is greater than one. If the
length of files is equal to one, then ignored. If FALSE then a list is returned.}
}
\description{
This function takes in sea level data and puts it into a form
that allows for easy use for calculations and data output. It's similar in nature
to as.ahccd.
}
\author{
Chantelle Layton
}
