% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscInstruments.R
\name{read.sealogT}
\alias{read.sealogT}
\title{Read sealog-T file}
\usage{
read.sealogT(file, skipStart = NA, skipEnd = NA, pressure)
}
\arguments{
\item{file}{a connection or character string  of the file to read.}

\item{skipStart}{an optional numeric value indicating number of hours to skip at the beginning of the file,
default is NA, meaning keep all the data points.}

\item{skipEnd}{an optional numeric value indicating number of hours to skip at the beginning of the file,
default is NA, meaning keep all the data points.}

\item{pressure}{a numerical value of the pressure that the instrument was moored}
}
\value{
a ctd object with \code{deploymentType = "moored"}.
}
\description{
This function reads in a sealog-T file.
}
\details{
Reads a sealog-T file, calculates the date based on the start, end, and sample period specified
in the header lines. Saves no information. It assumes that only temperature is present. This code
might be fragile, so use with caution.
}
\author{
Chantelle Layton
}
