% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscInstruments.R
\name{read.stAndrewsTemperature}
\alias{read.stAndrewsTemperature}
\title{Read non-archived St.Andrews moored temperature data}
\usage{
read.stAndrewsTemperature(file, skip = 0, header = TRUE, pressure)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load.}

\item{skip}{a numerical value indicating the number of rows to skip before reading in data}

\item{header}{logical value indicating whether or not a header is present in the file}

\item{pressure}{a numerical value of the pressure that the instrument was moored}
}
\value{
a ctd object with \code{deploymentType = "moored"}.
}
\description{
Read non-archived St.Andrews moored temperature data
}
\details{
This function attempts to read in a non-archived file that is obtained
from a technician at the St.Andrews Biological Station in St.Andrews, New Brunswick.
This is in no way a standard formatted file, as they are recieved in various
formats year to year. Note that pressure must be indicated as the files recieved have
no pressure information.
}
\author{
Chantelle Layton
}
