% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosewayPolygon}
\alias{rosewayPolygon}
\title{Roseway polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
rosewayPolygon
}
\description{
Polygon defining the bounds to look for data along the ancillary Atlantic Zone Monitoring Program
line, Roseway. It was constructed by finding the angle between the first point and the last point
of the line, and then calculating the bounding box which is 8 km wide. Stations include
RL_00, RL_02 : RL_08, RL_01 is omitted as it is not along the transect.
Note that nominal station locations could not be located, therefore, locations were pulled when the full
transect was occupied during HUD2001009.
}
\keyword{datasets}
