% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosewayStationPolygons}
\alias{rosewayStationPolygons}
\title{Roseway station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
rosewayStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the ancillary Atlantic Zone Monitoring Program
line, Roseway. Stations include RL_00, RL_02 : RL_08, RL_01 is omitted as it is not along the transect.
Note that nominal station locations could not be located, therefore, locations were pulled when the full
transect was occupied during HUD2001009.
Defining bounds are determined by station spacing. If the spacing is less than 8km,
then it is modified to half of the distance between the adjacent spacing. Note that this mean that the spacing
on either side can be uneven.
}
\keyword{datasets}
