% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{rotateAxes}
\alias{rotateAxes}
\title{Rotate axes}
\usage{
rotateAxes(x, y, angle)
}
\arguments{
\item{x}{a numeric vector of the x-component value(s)}

\item{y}{a numeric vector of the y-component value(s)}

\item{angle}{rotation angle in degrees}
}
\value{
list of rotated x and y components as \code{X} and \code{Y}.
}
\description{
Rotate axes
}
\details{
This function rotates points counterclockwise in a x-y Cartesian coordinate system.
To do clockwise rotation, simple supply a negative angle value.
}
\references{
\enumerate{
\item Basic explanation for rotation of axes \url{https://en.wikipedia.org/wiki/Rotation_of_axes}
}
}
\author{
Chantelle Layton
}
