% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{rotateAxesTrueNorth}
\alias{rotateAxesTrueNorth}
\title{Rotate axes clockwise from true north}
\usage{
rotateAxesTrueNorth(x, y, angle)
}
\arguments{
\item{x}{a numeric vector of the x-component value(s)}

\item{y}{a numeric vector of the y-component value(s)}

\item{angle}{rotation angle in degrees}
}
\value{
list of rotated x and y components as \code{X} and \code{Y}.
}
\description{
Rotate axes clockwise from true north
}
\details{
This function rotates points clockwise from true north in a x-y Cartesian coordinate system.
Adapted from MATLAB code provided to me by David Hebert, which was used in Mathieu Dever's PhD.
}
\author{
Chantelle Layton
}
