% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sableIslandBankNewPolygon}
\alias{sableIslandBankNewPolygon}
\title{Sable Island Bank New polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
sableIslandBankNewPolygon
}
\description{
Polygon defining the bounds to look for data along the cross-shelf portion of the ancillary Atlantic Zone Monitoring Program
line, Sable Island Bank, that was introduced in spring 2022. It was constructed by finding the angle between the first
point and the last point of the line, and then calculating the bounding box which is 8 km wide.
}
\keyword{datasets}
