% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sableIslandBankPolygon}
\alias{sableIslandBankPolygon}
\title{Sable Island Bank polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
sableIslandBankPolygon
}
\description{
Polygon defining the bounds to look for data along the ancillary Atlantic Zone Monitoring Program
line, Sable Island Bank. It was constructed by finding the angle between the first point and the last point
of the line, and then calculating the bounding box which is 8 km wide. Note that nominal station coordinates
could not be located. They were pulled from a cruise report.
}
\keyword{datasets}
