% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sableIslandBankStationPolygons}
\alias{sableIslandBankStationPolygons}
\title{Sable Island Bank station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
sableIslandBankStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the ancillary Atlantic Zone Monitoring Program
line, Sable Island Bank. Defining bounds are determined by station spacing. Note that nominal station coordinates
could not be located. They were pulled from a cruise report. If the spacing is less than 8km,
then it is modified to half of the distance between the adjacent spacing. Note that this mean that the spacing
on either side can be uneven.
}
\keyword{datasets}
