% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sableIslandBankWestEastStationPolygons}
\alias{sableIslandBankWestEastStationPolygons}
\title{Sable Island Bank West-East station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
sableIslandBankWestEastStationPolygons
}
\description{
Polygons defining bounds to look for data at each station in the west-east direction on the shelf portion of the
ancillary Atlantic Zone Monitoring Program line, Sable Island Bank, that was introduced in spring 2022.
Defining bounds are determined by station spacing. If the spacing is less than 8km,
then it is modified to half of the distance between the adjacent spacing. Note that this mean that the spacing
on either side can be uneven.
}
\keyword{datasets}
