% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotianShelfLonLatGrid}
\alias{scotianShelfLonLatGrid}
\title{Scotian shelf longitude latitude grid.}
\format{a data frame with un-named columns, but they are as follows,
\code{latitude}, \code{longitude}, \code{depth}, ...}
\usage{
scotianShelfLonLatGrid
}
\description{
Scotian shelf longitude latitude grid used for optimal estimation analysis (oax)
}
\keyword{datasets}
