% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{shelfSurveyInterpBarnes}
\alias{shelfSurveyInterpBarnes}
\title{Shelf survey barnes interpolation method}
\usage{
shelfSurveyInterpBarnes(
  ctd,
  fullgrid,
  grid,
  variable,
  xg,
  yg,
  xr,
  yr,
  iterations,
  gamma,
  averagingMethod,
  trimOutput,
  trimDistance
)
}
\arguments{
\item{ctd}{a list of \linkS4class{ctd} objects}

\item{fullgrid}{a data.frame containing the full original grid which contains columns, \code{x} which is the longitude,
\code{y} which is the latitude, \code{depth}, and \code{isBottom} which indicates if the grid point is the bottom depth.}

\item{grid}{a data.frame containing the grid points that data should be interpolated. This includes columns
\code{y} which is the latitude, \code{x}, which is the longitude, \code{depth}, a positive value, \code{tolerance}, a positive value
indicating a depth range used to get an average value at the given depth, \code{isBottom}, an optional column
indicating if the grid point is a bottom depth point.}

\item{variable}{a character value specifying which variable to extract from the \code{ctd} profiles.}

\item{xg, yg}{vectors defining x and y grids, see \link{interpBarnes} for additional clarification.}

\item{xr, yr}{lengths of the x and y grids for barnes interpolation, see \link{interpBarnes} for additional clarification.}

\item{iterations}{number of iterations for barnes interpolation, see \link{interpBarnes} for additional clarification.}

\item{gamma}{grid-focussing parameter for barnes interpolation, see \link{interpBarnes} for additional clarification.}

\item{averagingMethod}{a character vector of length one indicating how to vertically average the data to input
into the barnes interpolation. Current options include \code{mean} and \code{triangleWeight}.}

\item{trimOutput}{a logical value indicating if the barnes output should trim the grid output based on the input
data. If \code{TRUE}, any grid points which don't have any data within the value supplied to \code{trimDistance}.}

\item{trimDistance}{a numeric value indicating the distance, in kilometers, where barnes interpolation
output will be deemed valid if supplied data falls within the given distance. If not, the barnes output
at that grid point will be assigned a value of \code{NA}.}
}
\value{
a data.frame of the barnes interpolation output at each grid point defined in full grid, if the given
depth value and or grid points are specified in grid. The data.frame includes \code{x} the longitude, \code{y} the latitude,
\code{depth} the depth, and \code{data} the barnes output of the smoothed interpolated data for the given variable.
}
\description{
This function applies the appropriate methodology decided on for interpolating shelf survey
CTD measurements that has been tailored to the scotian shelf.
}
\author{
Chantelle Layton
}
