% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snowCrabOaxResults}
\alias{snowCrabOaxResults}
\title{Snow crab optimal estimation results}
\format{a data frame with columns \code{year}, \code{latitude}, \code{longitude},
\code{depth}, \code{temperature}, \code{temperature_error}, \code{salinity},
\code{salinity_error}}
\usage{
snowCrabOaxResults
}
\description{
Snow crab optimal estimation results from 1970 to present
}
\keyword{datasets}
