% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snowCrabUtmGrid}
\alias{snowCrabUtmGrid}
\title{Snow crab utm grid.}
\format{
a data frame with un-named columns, but they are as follows,
\code{latitude}, \code{longitude}, \code{depth}, ...
}
\usage{
snowCrabUtmGrid
}
\description{
Snow crab utm grid used for optimal estimation analysis (oax)
}
\keyword{datasets}
