% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sstInSituHalifax}
\alias{sstInSituHalifax}
\title{Sea surface temperature in-situ for Halifax site}
\format{a data frame with columns \code{year}, \code{month}, \code{temperature}}
\usage{
sstInSituHalifax
}
\description{
Monthly average sea surface temperature for in-situ Halifax site
that is located at the wharf at theMaritime museum in the Halifax harbour. The monthly averages
are only up to a certain year, the purpose of this file being included is that the
historical data is not avaliable anywhere.
}
\keyword{datasets}
