% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sstInSituPolygons}
\alias{sstInSituPolygons}
\title{Sea surface temperature in-situ polygons for two sites}
\format{
a list of length two, both of which have the station name and the longitude and latitude
}
\usage{
sstInSituPolygons
}
\description{
Longitude and latitude polygon definitions that were manually created to help facilitate
finding data.
}
\keyword{datasets}
