% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sstInSituStAndrews}
\alias{sstInSituStAndrews}
\title{Sea surface temperature in-situ for St.Andrews site}
\format{a data frame with columns \code{year}, \code{month}, \code{temperature}}
\usage{
sstInSituStAndrews
}
\description{
Monthly average sea surface temperature for in-situ St.Andrews site
that is located off the wharf at the St.Andrews Biological Station. The monthly averages
are only up to a certain year, the purpose of this file being included is that the
historical data is not avaliable anywhere.
}
\keyword{datasets}
