% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sstSatelliteNafoZones}
\alias{sstSatelliteNafoZones}
\title{Sea-surface temperature NAFO zones for Maritimes regions}
\format{
a list for each NAFO zone, where each zone has \code{zoneName}, \code{longitude}, and
\code{latitude}
}
\usage{
sstSatelliteNafoZones
}
\description{
NAFO zone definitions used for the zonal sea-surface temperature analysis, those only concerned with in the Maritimes
region are included here. The definitions were provided by P.Galbraith on 2022 07 13.
}
\keyword{datasets}
