% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stAnnsBankPolygon}
\alias{stAnnsBankPolygon}
\title{St.Anns Bank polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
stAnnsBankPolygon
}
\description{
Polygon defining the bounds to look for data along the core Atlantic Zone Monitoring Program
line, St.Anns Bank. It was constructed by finding the angle between the first point and the last point
of the line, and then calculating the bounding box which is 8 km wide.
}
\keyword{datasets}
