% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stAnnsBankStationPolygons}
\alias{stAnnsBankStationPolygons}
\title{St.Anns Bank station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
stAnnsBankStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the core Atlantic Zone Monitoring Program
line, St.Anns Bank.
}
\keyword{datasets}
