% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station2Climatology1980to2010}
\alias{station2Climatology1980to2010}
\title{Station 2 climatology}
\format{
a data frame with columns \code{month}, \code{depth},
\code{temperature_ltm}, \code{salinity_ltm}, \code{sigmaTheta_ltm}
}
\usage{
station2Climatology1980to2010
}
\description{
Station 2 1980 to 2010 climatology
}
\keyword{datasets}
