% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station2DepthBins}
\alias{station2DepthBins}
\title{Station 2 standard depth bins}
\format{
a data frame with columns \code{bin} and \code{tolerance}
}
\usage{
station2DepthBins
}
\description{
Station 2 depth bins and tolerances definitions for vertically averaging
}
\keyword{datasets}
