% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station2monthlyData}
\alias{station2monthlyData}
\title{Station 2 monthly data}
\format{a data frame with columns \code{year}, \code{month}, \code{depth},
\code{temperature}, \code{salinity}, \code{sigmaTheta}}
\usage{
station2monthlyData
}
\description{
Station 2 monthly and vertically averaged data for the entire timeseries.
}
\keyword{datasets}
