% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station2PlotLimits}
\alias{station2PlotLimits}
\title{Station plot limits}
\format{
a list of \code{limits}, \code{contourLevelLimits}, and \code{contourLevelLimits},
each having defined items for \code{temperature}, \code{temperatureAnomaly}, \code{salinity},
\code{salinityAnomaly}, \code{sigmaTheta}, \code{sigmaThetaAnomaly}.
}
\usage{
station2PlotLimits
}
\description{
Various limits for plotting station 2 section plots.
}
\keyword{datasets}
