% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station2Polygon}
\alias{station2Polygon}
\title{Station 2 polygon}
\format{
a list containing \code{longitude} and \code{latitude}
}
\usage{
station2Polygon
}
\description{
Polygon defining the bounds to look for data along the fixed Atlantic Zone Monitoring Program
station, Station 2. It was carried over from previous employee.
}
\keyword{datasets}
