% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishCilGrid}
\alias{summerGroundfishCilGrid}
\title{Summer groundfish cold intermediate layer (cil) grid.}
\format{
a data frame \code{latitude} and \code{longitude}.
}
\usage{
summerGroundfishCilGrid
}
\description{
Summer groundfish cold intermediate layer (cil) grid, it is a subset of the summerGroundfishLonLatGrid
}
\keyword{datasets}
