% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishClimatology1980to2010}
\alias{summerGroundfishClimatology1980to2010}
\title{Summer groundfish optimal estimation climatology}
\format{a data frame with columns \code{latitude}, \code{longitude}, \code{depth},
\code{temperature}, \code{temperature_error}, \code{salinity}, \code{salinity_error},
\code{sigmat}, \code{sigmat_error}.}
\usage{
summerGroundfishClimatology1980to2010
}
\description{
Summer groundfish optimal estimation climatology for 1980 to 2010.
}
\keyword{datasets}
