% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishOaxNafoPolygons}
\alias{summerGroundfishOaxNafoPolygons}
\title{Summer groundfish NAFO region boundaries for oax grid}
\format{
a list that contains each NAFO region within the oax grid, 4vn, 4vs,
4w, 4x, and 4v. Each list element contains \code{longitude} and \code{latitude}.
}
\usage{
summerGroundfishOaxNafoPolygons
}
\description{
Summer groundfish NAFO region boundaries based on the grid that is used for the
OAX analysis.
}
\keyword{datasets}
