% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishOaxResultsForCIL}
\alias{summerGroundfishOaxResultsForCIL}
\title{Summer groundfish optimal estimation results to use for cold intermediate layer calculations}
\format{
a data frame with columns \code{year}, \code{latitude}, \code{longitude},
\code{depth}, \code{temperature}, \code{salinity},
}
\usage{
summerGroundfishOaxResultsForCIL
}
\description{
Summer groundfish optimal estimation results from 1972 to present. This is different than
summerGroundfishOaxResults as the analysis was done differently. Due to low vertical resolution
in the earlier years, data was combined to get better results for e.g 1972 includes data from 1970
to 1974. Data from 1990 to 2017 is all high resolution data, so it is only the data from that year.
This data is already subsetted to the CIL grid, except for 2019, which is the full OAX grid.
}
\keyword{datasets}
