% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishStandardDepths}
\alias{summerGroundfishStandardDepths}
\title{Summer groundfish standard grid depths.}
\format{
a numeric vector
}
\usage{
summerGroundfishStandardDepths
}
\description{
Summer groundfish standard grid depths used in summerGroundfishGrid
}
\keyword{datasets}
