% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishUtmGrid}
\alias{summerGroundfishUtmGrid}
\title{Summer groundfish utm grid.}
\format{
a data frame with un-named columns, but they are as follows,
\code{latitude}, \code{longitude}, \code{depth}, ...
}
\usage{
summerGroundfishUtmGrid
}
\description{
Summer groundfish utm grid used for optimal estimation analysis (oax)
}
\keyword{datasets}
