% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishXg}
\alias{summerGroundfishXg}
\title{Summer groundfish xg for input into barnes interpolation.}
\format{
a numeric vector
}
\usage{
summerGroundfishXg
}
\description{
Summer groundfish standard x-grid points used in summerGroundfishGrid
}
\keyword{datasets}
