% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerGroundfishYg}
\alias{summerGroundfishYg}
\title{Summer groundfish yg for input into barnes interpolation.}
\format{
a numeric vector
}
\usage{
summerGroundfishYg
}
\description{
Summer groundfish standard y-grid points used in summerGroundfishGrid
}
\keyword{datasets}
