% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw.R
\name{swSatO2}
\alias{swSatO2}
\title{Saturation of O2 in sea water}
\usage{
swSatO2(temperature, salinity)
}
\arguments{
\item{temperature}{a numeric vector of temperature measurements}

\item{salinity}{a numeric vector of salinity measurements}
}
\description{
Saturation of O2 in sea water
}
\details{
Computes the solubility (saturation) of oxygen in sea water. Based on the
matlab function SW_SATO2 from the CSIRO seawater toolbox.
}
\author{
Chantelle Layton
}
