% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theGullyNewPolygon}
\alias{theGullyNewPolygon}
\title{The Gully polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
theGullyNewPolygon
}
\description{
Polygon defining the bounds to look for data along the Atlantic Zone Monitoring Program
sampling area, The Gully. After a report completed in 2021, new stations were decided on.
Therefore this is referred to as the "new" Gully.
}
\keyword{datasets}
