% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theGullyNewStationPolygons}
\alias{theGullyNewStationPolygons}
\title{The Gully station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
theGullyNewStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the Atlantic Zone Monitoring Program
sampling area, The Gully. It was constructed by finding the angle between the first two stations,
and the last two stations, and then calculating the bounding box which is 8 km wide.
}
\keyword{datasets}
