% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theGullyPolygon}
\alias{theGullyPolygon}
\title{The Gully polygon}
\format{
a list of \code{longitude} and \code{latitude}
}
\usage{
theGullyPolygon
}
\description{
Polygon defining the bounds to look for data along the Atlantic Zone Monitoring Program
sampling area, The Gully. It was constructed by calculating the angle between the head and mouth stations,
and then the two shelf stations, resulting in a 'T' shaped polygon.
}
\keyword{datasets}
