% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theGullyStationPolygons}
\alias{theGullyStationPolygons}
\title{The Gully  station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
theGullyStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the Atlantic Zone Monitoring Program
sampling area, The Gully. Due to the unique shape of the sampling done in The Gully, a box 8km by 7.6km was
constructed around each station. For the head and mouth stations, GULD_03 and GULD_04 respectively, the angle
between those two stations was used to construct the bounding box. For the two shelf stations, SG_23 and SG_28,
the angle between those two stations was used to construct the bounding box.
}
\keyword{datasets}
