% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transectClimatology1980to2010}
\alias{transectClimatology1980to2010}
\title{Transect 1980 to 2010 climatology}
\format{a list which contains a list for each transect that has the \code{transecAbbreviation},
\code{transectFullname}, \code{season}, \code{program}, and \code{climatology}, the climatology has
column names of \code{distanceBin}, \code{depthBin}, \code{temperature}, \code{temperatureSd},
\code{temperatureNobs}, \code{salinity}, \code{salinitySd}, \code{salinityNobs}, \code{sigmaTheta},
\code{sigmaThetaSd}, \code{sigmaThetaNobs}.}
\usage{
transectClimatology1980to2010
}
\description{
Core azmp transect 1980 to 2010 climatology.
}
\keyword{datasets}
