% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transectDefinitions}
\alias{transectDefinitions}
\title{Transect definitions}
\format{a list which contains a list for each transect that has \code{info}, that has a handful
of other variables which is too lengthy to list here, but two that are of importance are
\code{start_longitude} and \code{start_latitude}, \code{bottom_outline}, the bathymetry for plotting,
and then other information that is not currently used, \code{bottom_mask}, \code{xgrid}, and \code{ygrid}.}
\usage{
transectDefinitions
}
\description{
Transect definitions has some important information pertaining to plotting
many of the transects associated with the AZMP program. Some is leftover from what used to be used
when plotting in MATLAB, but there are some items of importance, so everything was saved.
}
\keyword{datasets}
