% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transectDepthBins}
\alias{transectDepthBins}
\title{Transect standard depth bins}
\format{
a data frame with columns \code{bin} and \code{tolerance}
}
\usage{
transectDepthBins
}
\description{
Transect standard depth bins and tolerances definitions for verticaly averaging
}
\keyword{datasets}
