% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transectDistanceBins}
\alias{transectDistanceBins}
\title{Transect distance bins}
\format{
a list with the transect name, each with a data frame with columns \code{bin} and
\code{tolerance}. Only one transect has separate definitions based on the season.
}
\usage{
transectDistanceBins
}
\description{
Transect distance bin for the 1981 to 2010 climatology definitions for horizontal averaging
}
\keyword{datasets}
