% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transectPlotLimits}
\alias{transectPlotLimits}
\title{Transect plot limits}
\format{
a list of \code{limits}, \code{contourLevelLimits}, and \code{contourLevelLimits},
each having defined items for \code{temperature}, \code{temperatureAnomaly}, \code{salinity},
\code{salinityAnomaly}, \code{sigmaTheta}, \code{sigmaThetaAnomaly}.
}
\usage{
transectPlotLimits
}
\description{
Various limits for plotting transects.
}
\keyword{datasets}
