% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{transectPoly}
\alias{transectPoly}
\title{Make a transect polygon}
\usage{
transectPoly(x, y, yadj = 50)
}
\arguments{
\item{x, y}{vectors containg the coordinates of the verticies of the polygon}

\item{yadj}{a numeric value indicating the amount to increase the maximum \code{y} value.}
}
\value{
a list of \code{xpoly} and \code{ypoly} values
}
\description{
This function takes coordinates for a polygon and will complete it for plotting purposes by
extending the limit of the bottom of the polygon by some numeric amount.
}
\author{
Chantelle Layton
}
