% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{weightedArea}
\alias{weightedArea}
\title{Calculate weighted area for binned data}
\usage{
weightedArea(bins, binarea)
}
\arguments{
\item{bins}{a numeric vector indicating the bins}

\item{binarea}{a numeric vector of the total area calculated for the specified bin}
}
\description{
Calculate weighted area value for binned data
}
\author{
Chantelle Layton
}
