% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishGrid}
\alias{winterGroundfishGrid}
\title{Winter groundfish grid.}
\format{
a data frame columns \code{x}, or the longitude, \code{y}, or the latitude,
\code{depth}, \code{tolerance}, and \code{isBottom} which is a logical column indicating
if the point is considered a bottom depth of the grid.
}
\usage{
winterGroundfishGrid
}
\description{
Winter groundfish grid used interpolation
}
\keyword{datasets}
