% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishGridOld}
\alias{winterGroundfishGridOld}
\title{Winter groundfish grid.}
\format{
a data frame columns \code{x}, or the longitude, \code{y}, or the latitude,
\code{depth}, \code{tolerance}, and \code{isBottom} which is a logical column indicating
if the point is considered a bottom depth of the grid.
}
\usage{
winterGroundfishGridOld
}
\description{
Winter groundfish grid used for interpolation, note that this is the depreciated grid, but
is kept for completeness, it may be removed in the future.
}
\keyword{datasets}
