% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishOaxCtdInput}
\alias{winterGroundfishOaxCtdInput}
\title{Winter groundfish optimal estimation ctd input}
\format{
a list of ctd objects
}
\usage{
winterGroundfishOaxCtdInput
}
\description{
Winter groundfish optimal estimation ctd input from 2008 to present. The data was put
input ctd objects from the input files for use of plotting multiple years of oax results
}
\keyword{datasets}
