% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishOaxResults}
\alias{winterGroundfishOaxResults}
\title{Winter groundfish optimal estimation results}
\format{a data frame with columns \code{year}, \code{latitude}, \code{longitude},
\code{day_number}, \code{depth}, \code{temperature}, \code{temperature_error}, \code{salinity},
\code{salinity_error}}
\usage{
winterGroundfishOaxResults
}
\description{
Winter groundfish optimal estimation results from 1970 to present
}
\keyword{datasets}
