% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishPlotLimits}
\alias{winterGroundfishPlotLimits}
\title{Winter groundfish map plot limits}
\format{a list of \code{limits}, \code{contourLevelLimits}, and \code{contourLevelLimits},
each having defined items for \code{temperature}, \code{temperatureAnomaly}, \code{salinity},
and \code{salinityAnomaly}.}
\usage{
winterGroundfishPlotLimits
}
\description{
Various limits for plotting summer groundfish maps.
}
\keyword{datasets}
