% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishStandardDepths}
\alias{winterGroundfishStandardDepths}
\title{Winter groundfish standard grid depths.}
\format{
a numeric vector
}
\usage{
winterGroundfishStandardDepths
}
\description{
Winter groundfish standard grid depths used in winterGroundfishGrid
}
\keyword{datasets}
