% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishXg}
\alias{winterGroundfishXg}
\title{Winter groundfish xg for input into barnes interpolation.}
\format{
a numeric vector
}
\usage{
winterGroundfishXg
}
\description{
Winter groundfish standard x-grid points used in winterGroundfishGrid
}
\keyword{datasets}
