% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterGroundfishYg}
\alias{winterGroundfishYg}
\title{Winter groundfish yg for input into barnes interpolation.}
\format{
a numeric vector
}
\usage{
winterGroundfishYg
}
\description{
Winter groundfish standard y-grid points used in winterGroundfishGrid
}
\keyword{datasets}
