% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{writeProductData}
\alias{writeProductData}
\title{Write data product output}
\usage{
writeProductData(
  filename,
  destdir = ".",
  dateCreated = presentTime(),
  contactName,
  contactEMail,
  description = NULL,
  otherMetadata = NULL,
  data,
  ndigits = 1
)
}
\arguments{
\item{filename}{a character string indicating the name of the file.}

\item{destdir}{Optional string indicating the directory in which to store downloaded files.
If not supplied, "." is used, i.e. the data file is stored in the present working directory.}

\item{dateCreated}{a POSIX format date indicating the time of creation, the
default is set using the \code{presentTime}.}

\item{contactName}{a character string indicating the person of contact}

\item{contactEMail}{a character string indicating the contact name's e-mail address.}

\item{description}{an optional character string providing some details of the data product.}

\item{otherMetadata}{an optional list indicating additional metadata to provide in the
header of the file. The list names should be provided in camelCase.}

\item{data}{a data.frame of the data that is to be put into the file with appropriate
column names, again, prefferably in camelCase}

\item{ndigits}{the number of digits to output the data, default is 2}
}
\description{
This function will output a file that is used for dissemination
of the data products produced for the research document.
}
\author{
Chantelle Layton
}
