% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{annualAnomaly}
\alias{annualAnomaly}
\title{Calculate annual anomaly values}
\usage{
annualAnomaly(d, minNMonths = 12)
}
\arguments{
\item{d}{a data.frame containing year, month, and at least anomaly, and optionally normalizedAnomaly}

\item{minNMonths}{an numeric value indicating the minimum number of months required to calculate an
annual value for a given year. The default value is set to 12.}
}
\value{
a data.frame with year, anomaly, and optionally normalizedAnomaly
}
\description{
This function calculates annual anomaly values from provided
monthly anomaly values.
}
\author{
Chantelle Layton
}
