% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{averageCtds}
\alias{averageCtds}
\title{Average a list of CTD objects.}
\usage{
averageCtds(x)
}
\arguments{
\item{x}{a list of ctd objects}
}
\value{
a ctd object that has the same metadata and processing log as the supplied ctd object,
but with bin-averaged data.
}
\description{
Average and list of CTD objects. This function will take all available data within all of the ctd objects
and will average in pressure space. It takes all unique pressure values from all ctd objects, and then averages
all availble data for each unique pressure values. Therefore, it's best if the data have already been vertically averaged.
}
\author{
Chantelle Layton
}
