% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculateMixedLayerDepth}
\alias{calculateMixedLayerDepth}
\title{Calculate mixed layer depth}
\usage{
calculateMixedLayerDepth(
  ctd,
  method = "gradient",
  mldMin = NULL,
  nPointsMin = 3,
  densityThreshold = 0.03,
  densityReferenceDepth = 10,
  debug = TRUE
)
}
\arguments{
\item{ctd}{a \code{ctd} object.}

\item{method}{a character string indicting which method to
use to calculate the mixed layer depth, options include
'gradient', 'N2', and 'density'. Default is 'gradient'}

\item{mldMin}{a numeric value indicating the minimum value the
mixed layer depth should be. Default is set to \code{NULL}.
Ignored for \code{method = 'density'}.}

\item{nPointsMin}{a numeric value indicating the minimum number of
points in ctd object}

\item{densityThreshold}{a numeric value indicating the threshold value
to use for the 'density' method, ignored for other methods.
Default is 0.03.}

\item{densityReferenceDepth}{a numeric value indicating the reference
depth for the density method. It will find the closest
value in the profile within 5dbar. Default is 10dbar.}

\item{debug}{a logical value indicating to show debug information}
}
\value{
a single numeric value of the calculated mixed layer depth
}
\description{
This function calculated the mixed layer depth, which is
defined as when the density gradient, or d(sigmaTheta)/dz, is
greater than 0.01 kg/m^-4. A parameter, \code{mldMin}, can be provided
in order to force the mixed layer depth to be found below the
set depth.
}
\author{
Chantelle Layton and Benoit Casault
}
