% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculateMixedLayerDepth}
\alias{calculateMixedLayerDepth}
\title{Calculate mixed layer depth}
\usage{
calculateMixedLayerDepth(ctd, mldMin = NULL, debug = TRUE)
}
\arguments{
\item{ctd}{a \code{ctd} object.}

\item{mldMin}{a numeric value indicating the minimum value the
mixed layer depth should be. Default is set to \code{NULL}}

\item{debug}{a logical value indicating to show debug information}
}
\value{
a single row data frame with relevant information for easy output use
}
\description{
This function calculated the mixed layer depth, which is
defined as when the density gradient, or d(sigmaTheta)/dz, is
greater than 0.01 kg/m^-4. A parameter, \code{mldMin}, can be provided
in order to force the mixed layer depth to be found below the
set depth.
}
\author{
Chantelle Layton and Benoit Casault
}
