% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laurentianChannelMouthStationPolygons}
\alias{laurentianChannelMouthStationPolygons}
\title{Laurentian Channel Mouth station polygon's}
\format{
a list which contains \code{stationName}, \code{longitude}, \code{latitude}, \code{polyLongitude},
\code{polyLatitude} for each station.
}
\usage{
laurentianChannelMouthStationPolygons
}
\description{
Polygons defining bounds to look for data at each station along the ancillary Atlantic Zone Monitoring Program
line, Laurentian Channel Mouth. Defining bounds are determined by station spacing. If the spacing is less than 8km,
then it is modified to half of the distance between the adjacent spacing. Note that this mean that the spacing
on either side can be uneven.
}
\keyword{datasets}
