% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{monthlyClimatology}
\alias{monthlyClimatology}
\title{Calculate monthly climatology}
\usage{
monthlyClimatology(d, climatologyYears)
}
\arguments{
\item{d}{a data.frame containing at least a column named month and one other variable}

\item{climatologyYears}{a vector of length two indicating the range of years
to calculate the climatology}
}
\value{
the results of aggregate
}
\description{
This function calculates the monthly climatology
}
\author{
Chantelle Layton
}
