% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMonthlyTimeseriesWAnomalyBar}
\alias{plotMonthlyTimeseriesWAnomalyBar}
\title{Plot monthly timeseries with anomaly bar}
\usage{
plotMonthlyTimeseriesWAnomalyBar(
  xYear,
  xMonth,
  y,
  yAnomaly,
  yNormalizedAnomaly,
  xAnnualAnomaly,
  yAnnualAnomaly,
  plotAnnual = TRUE,
  xClimatology,
  yClimatology,
  sdClimatology,
  xlim,
  ylim,
  anomalyColors,
  scorecardSeparation = par("cin")[2]/2 + 0.1 - 0.05 - 0.0125,
  scorecardWidth = par("cin")[2] + 0.2 - 0.05 - 0.0125
)
}
\arguments{
\item{xYear}{a numeric vector of the associated year}

\item{xMonth}{a numeric vector of the associated month}

\item{y}{a numeric vector of the associated year month value}

\item{yAnomaly}{a numeric vector of the associated anomaly year month value}

\item{yNormalizedAnomaly}{a numeric vector of the associated normalized anomaly with xYear and xMonth}

\item{xAnnualAnomaly}{a numeric vector of the year for the annual anomaly}

\item{yAnnualAnomaly}{a numeric vector of the annual anomaly}

\item{plotAnnual}{a logical value indicating if the annual anomaly should be shown in the scorecard bar, default is \code{TRUE}}

\item{xClimatology}{a numeric vector of the months for the monthly climatology}

\item{yClimatology}{a numeric vector of the climatology values for associated month}

\item{sdClimatology}{a numeric vector of the standard deviation for the associated month}

\item{xlim}{the x limits (x1, x2) of the plot}

\item{ylim}{the y limits (y1, y2) of the plot}

\item{anomalyColors}{a list that contains the colors and breaks of the desired color bar for the anomaly scorecard}

\item{scorecardSeparation}{Value used to denote the separation between the main plot and scorecard.
Suggested to use \code{par('cin')[2]} as a starting point and play with that value. If closer is desired, decrease.}

\item{scorecardWidth}{Value used to denote the width of the scorecard. Suggested to use
\code{par('cin')[2]} as a starting point and play with that value. If wider is desired, increase.}
}
\description{
Creates a time series plot with the monthly anomaly and annual
anomaly scorecard below.
}
\details{
This is the plot style that is used in many of Peter Galbraith's
research documents. Was translated to R in efforts to better represent
some data.
}
\author{
Chantelle Layton
}
