% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotStackedBarplot}
\alias{plotStackedBarplot}
\title{Plot annual stacked bar plot}
\usage{
plotStackedBarplot(
  x,
  y,
  z,
  plotAverage = TRUE,
  ylab1 = NULL,
  ylab2 = NULL,
  ylim1,
  ylim2,
  ncol = 1
)
}
\arguments{
\item{x}{a vector indicating what the columns represent for \code{z}}

\item{y}{a vector indicating what the rows represent for \code{z}}

\item{z}{a matrix with the columns being the x-axis and the rows being items
that wish to be stacked.}

\item{plotAverage}{a logical value indicating whether or not to add the average value
from matrix z.}

\item{ylab1}{name for the y-axis on side 2, default is \code{NULL}}

\item{ylab2}{name for the y-axis on side 4 if \code{plotAverage = TRUE}, ignored otherwise, default is
\code{NULL}.}

\item{ylim1}{limits for the y-axis on side 2, if not supplied, it will be inferred from \code{z}}

\item{ylim2}{limits for the y-axis on side 4 if \code{plotAverage = TRUE}, ignored otherwise. If not
supplied, it will be inferred from the data.}

\item{ncol}{supplied to legend, the number of columns in which to set the legend items, default is 1.}
}
\description{
This function will plot a stacked bar plot that has negative and
positive values. It also has the option to plot the average of values provided.
}
\author{
Chantelle Layton
}
