% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{rotateAxes}
\alias{rotateAxes}
\title{Rotate axes}
\usage{
rotateAxes(x, y, angle, inverse = FALSE)
}
\arguments{
\item{x}{a numeric vector of the x-component value(s)}

\item{y}{a numeric vector of the y-component value(s)}

\item{angle}{rotation angle in degrees}

\item{inverse}{logical value indicating whether or not to do inverse transformation, default is \code{FALSE}.}
}
\value{
list of rotated x and y components as \code{X} and \code{Y}.
}
\description{
Rotate axes
}
\details{
This function rotates points in a x-y Cartesian coordinate system. Setting \code{inverse = TRUE} is effectively the
same as multiplying the provided \code{angle} by \code{-1}, so for example, providing parameters \code{angle = 50} and \code{inverse = TRUE} would give
the same results providing parameters \code{angle = -50} and \code{inverse = FALSE}.
}
\references{
\enumerate{
\item Basic explanation for rotation of axes \url{https://en.wikipedia.org/wiki/Rotation_of_axes}
}
}
\author{
Chantelle Layton
}
